[LibOS] Move fully to inodes
This change removes the `state` dentry field, and uses the `inode` field
directly: a dentry is either positive (with inode) or negative (without
inode). Dentry fields describing the file (`type`, `perm`) are also
gone.

As a consequence, negative lookups are no longer cached: a negative
dentry only means that a file is not known to exist.

Signed-off-by: Paweł Marczewski <pawel@invisiblethingslab.com>